requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==1.10.0
pydantic==2.5.2
langchain==0.2.17
langchain-core==0.2.43
langchain-community>=0.0.20
langsmith==0.1.147
python-magic==0.4.27
pytube==15.0.0
moviepy==1.0.3
pydub==0.25.1
SpeechRecognition==3.10.0
flask==2.3.3
flask-socketio==5.3.6
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1 